requirements.txt
pytest
requests
allure-pytest
webdriver-manager